Add ability to limit write rate
For the purposes of load testing our logging infrastructure,
we would like to control how many bytes of logs are written
per second.

This commit creates a new abstraction which is then used
to control writes.  The abstraction tracks the number of bytes
uploaded over the last 10 seconds and adjusts how many bytes
should be written in the next batch to ensure the desire
write rate is not exceeded.

Also added support for a telemetry file.  The number of bytes
and lines created are written to this file once every second.
This allows us to cross reference the number of bytes received
with the number of bytes the generate thinks it wrote.